Completed the following --------------------------
1. index.js:
a. Refactored the code to use querySelector in the restartGameButtonClickEvent button event to remove the images for each img tag.
b. All the buttons have been made as global variables to reduce the number of DOM queries to search for elements.
c. Restart button: Made code changes to alter the behavior of the Restart button as follows
i) Reset button is disabled when the game is loaded for the first time.
ii) The restart button is disabled until another game is in progress or completed.
c. Hit and Stay buttons: The style property has been removed to be added as a CSS class instead.