Completed the following --------------------------
1. Index.js:
a. Implemented a callback function while creating image elements.
b. Refactored the code to use card collection arrays for player and dealer instead of individual variables and dynamically populate the card images using the method 'setCardImages'.
2. BJResult.js: Used switch-case statements instead of multiple if-else statements.
3. BJCardGenerator.js:
a. Implemented 'aceValueCollection' method to create an array of values associated with the Ace card instead of hardcoding.
b. Removed the unused method 'viewCardCollectionwithPoints'